[DOCS] Automate Jira connector screenshots (#165917)
(cherry picked from commit 87c60a381936314db9474921b3d1da033f276ae7)
