+++ docs/management/connectors/action-types/jira.asciidoc
@@ -25,6 +25,7 @@ or as needed when you're creating a rule. For example:
 
 [role="screenshot"]
 image::management/connectors/images/jira-connector.png[Jira connector]
+// NOTE: This is an autogenerated screenshot. Do not edit it directly.
 
 [float]
 [[jira-connector-configuration]]
@@ -80,6 +81,7 @@ as you're creating or editing the connector in {kib}. For example:
 
 [role="screenshot"]
 image::management/connectors/images/jira-params-test.png[Jira params test]
+// NOTE: This is an autogenerated screenshot. Do not edit it directly.
 
 Jira actions have the following configuration properties.
