+++ x-pack/test/screenshot_creation/apps/response_ops_docs/stack_connectors/pagerduty_connector.ts
@@ -13,35 +13,24 @@ export default function ({ getService, getPageObjects }: FtrProviderContext) {
   const pageObjects = getPageObjects(['common', 'header']);
   const actions = getService('actions');
   const testSubjects = getService('testSubjects');
-  const actionBody =
-    `{\n` +
-    `"model": "gpt-3.5-turbo",\n` +
-    `"messages": [{\n` +
-    `"role": "user",\n` +
-    `"content": "You are a cyber security analyst using Elastic Security. I would like you to evaluate the event below and format your output neatly in markdown syntax. Add your description, an accuracy rating, and a threat rating."\n` +
-    `}]`;
 
-  describe('generative ai connector', function () {
+  describe('pagerduty connector', function () {
     beforeEach(async () => {
       await pageObjects.common.navigateToApp('connectors');
       await pageObjects.header.waitUntilLoadingHasFinished();
     });
-
-    it('generative ai connector screenshots', async () => {
+    it('pagerduty connector screenshots', async () => {
       await pageObjects.common.navigateToApp('connectors');
       await pageObjects.header.waitUntilLoadingHasFinished();
-      await actions.common.openNewConnectorForm('gen-ai');
-      await testSubjects.setValue('nameInput', 'OpenAI test connector');
-      await testSubjects.setValue('secrets.apiKey-input', 'testkey');
-      await commonScreenshots.takeScreenshot('gen-ai-connector', screenshotDirectories, 1920, 1200);
+      await actions.common.openNewConnectorForm('pagerduty');
+      await testSubjects.setValue('nameInput', 'PagerDuty test connector');
+      await testSubjects.setValue('pagerdutyApiUrlInput', 'https://dev-test.pagerduty.com/');
+      await testSubjects.setValue('pagerdutyRoutingKeyInput', 'testkey');
+      await commonScreenshots.takeScreenshot('pagerduty-connector', screenshotDirectories);
       await testSubjects.click('create-connector-flyout-save-test-btn');
       await testSubjects.click('toastCloseButton');
-      const editor = await testSubjects.find('kibanaCodeEditor');
-      await editor.clearValue();
-      await testSubjects.setValue('kibanaCodeEditor', actionBody, {
-        clearWithKeyboard: true,
-      });
-      await commonScreenshots.takeScreenshot('gen-ai-params-test', screenshotDirectories);
+      await testSubjects.setValue('eventActionSelect', 'trigger');
+      await commonScreenshots.takeScreenshot('pagerduty-params-test', screenshotDirectories);
       await testSubjects.click('euiFlyoutCloseButton');
     });
   });
